requirements.txt
apex==0.1
backcall==0.1.0
certifi==2019.9.11
cffi==1.12.3
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
imageio==2.5.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
kiwisolver==1.1.0
-e git+https://github.com/facebookresearch/maskrcnn-benchmark.git@cd128b8ba3ce7ef09485231a44714b3ee11da7eb#egg=maskrcnn_benchmark
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.3
ninja==1.9.0.post1
numpy==1.17.2
olefile==0.46
opencv-python==4.1.1.26
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycocotools==2.0
pycparser==2.19
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.2
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
torch==1.1.0
torchvision==0.2.2
tqdm==4.36.1
traitlets==4.3.2
wcwidth==0.1.7
yacs==0.1.6
